chore(node/service): cleanup standard rollup node dir
